add `_Manager(ABC)` as a type
